pyproject.toml
[project]
name = "fastapi-saas-starter"
version = "0.1.0"
description = "Multi-tenant FastAPI SaaS starter with Temporal workflows"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "temporalio>=1.7.1",
    "httpx>=0.28.0",
    "slowapi>=0.1.9",
    "asgi-correlation-id>=4.0.0",
    "zxcvbn>=4.4.28",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "structlog>=24.0.0",
    "resend>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.8.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

# --- Ruff Configuration ---
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Enable specific rules:
# E/W: pycodestyle (standard errors/warnings)
# F: Pyflakes (logical errors)
# I: isort (import sorting)
# B: flake8-bugbear (common bugs)
# UP: pyupgrade (upgrades syntax, e.g., Set[] -> set())
# SIM: flake8-simplify (simplifications)
# FAST: FastAPI specific rules
select = ["E", "W", "F", "I", "B", "UP", "SIM", "FAST"]
# FAST001: response_model is often preferred for explicit documentation
ignore = ["FAST001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# --- MyPy Configuration ---
[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = false
plugins = [
    "pydantic.mypy",
]

# Exclude problematic directories
exclude = [
    "tests/",
    "src/alembic/",
]

# Per-module overrides for third-party libraries without type stubs
[[tool.mypy.overrides]]
module = [
    "temporalio.*",
    "slowapi.*",
    "asgi_correlation_id.*",
    "prometheus_fastapi_instrumentator.*",
    "jose",
    "jose.*",
    "zxcvbn",
    "zxcvbn.*",
    "argon2",
    "argon2.*",
    "resend",
    "resend.*",
]
ignore_missing_imports = true

# Allow untyped calls/decorators in modules that use libraries without stubs
[[tool.mypy.overrides]]
module = [
    "src.app.core.security",
    "src.app.core.logging",
    "src.app.schemas.auth",
    "src.app.temporal.activities",
    "src.app.temporal.workflows",
    "src.app.temporal.client",
    "src.app.api.v1.auth",
    "src.app.main",
]
disallow_untyped_calls = false
disallow_untyped_decorators = false
warn_return_any = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
